requirements.txt
google-generativeai>=0.3.0
streamlit>=1.28.0
bcrypt>=4.0.1
python-dotenv>=1.0.0
